gpm_airflow/requirements.txt
apache-airflow[celery,redis,postgres]==2.12.4
rasterio==1.3.9
xarray==2023.12.0
h5netcdf==1.2.0
netcdf4==1.6.5
h5py==3.10.0
numpy==1.24.4
matplotlib==3.7.4
pandas==2.1.4
geopandas==0.14.2
pyproj==3.6.1
shapely==2.0.2
fiona==1.9.5
scipy==1.11.4
seaborn==0.13.0
tqdm==4.66.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-dateutil==2.8.2
tzdata==2023.4
celery==5.3.4
redis==5.0.1
flower==2.0.1